Add ULongLong to list of types that needs a cast for sizeof

+++ tests/ClangSharp.PInvokeGenerator.UnitTests/Baseline/Baselines/SizeOfMacroTestTest/RuntimeSizeofMacroTest.CSharp.Latest.Windows.cs
@@ -9,6 +9,6 @@ namespace ClangSharp.Test
     public static unsafe partial class Methods
     {
         [NativeTypeName("#define XR_MAX_EVENT_DATA_SIZE sizeof(XrEventDataBuffer)")]
-        public static readonly ulong XR_MAX_EVENT_DATA_SIZE = sizeof(XrEventDataBuffer);
+        public static readonly ulong XR_MAX_EVENT_DATA_SIZE = (uint)(sizeof(XrEventDataBuffer));
     }
 }
